// mobile/src/api/types.ts
export interface LedgerConsumer {
  account_number: string;
  name: string | null;
  address: string | null;
  status: string | null;
  meter_no?: string | null;
  rate_class?: string | null;
}